+++ core/src/main/java/org/geysermc/geyser/entity/type/TridentEntity.java
@@ -33,7 +33,7 @@ import java.util.UUID;
 
 public class TridentEntity extends AbstractArrowEntity {
 
-    public TridentEntity(GeyserSession session, long entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
+    public TridentEntity(GeyserSession session, int entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
         super(session, entityId, geyserId, uuid, definition, position, motion, yaw, pitch, headYaw);
     }
 }
