+++ core/src/main/java/org/geysermc/geyser/entity/type/WitherSkullEntity.java
@@ -36,7 +36,7 @@ import java.util.UUID;
 public class WitherSkullEntity extends FireballEntity {
     private boolean isCharged;
 
-    public WitherSkullEntity(GeyserSession session, long entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
+    public WitherSkullEntity(GeyserSession session, int entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
         super(session, entityId, geyserId, uuid, definition, position, motion, yaw, pitch, headYaw);
 
         this.futureTicks = 1;
