+++ core/src/main/java/org/geysermc/geyser/entity/type/living/MobEntity.java
@@ -43,7 +43,7 @@ public class MobEntity extends LivingEntity {
     @Getter
     private long leashHolderBedrockId;
 
-    public MobEntity(GeyserSession session, long entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
+    public MobEntity(GeyserSession session, int entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
         super(session, entityId, geyserId, uuid, definition, position, motion, yaw, pitch, headYaw);
     }
