+++ core/src/main/java/org/geysermc/geyser/entity/type/living/animal/GoatEntity.java
@@ -41,7 +41,7 @@ public class GoatEntity extends AnimalEntity {
     @Getter
     private boolean isScreamer;
 
-    public GoatEntity(GeyserSession session, long entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
+    public GoatEntity(GeyserSession session, int entityId, long geyserId, UUID uuid, EntityDefinition<?> definition, Vector3f position, Vector3f motion, float yaw, float pitch, float headYaw) {
         super(session, entityId, geyserId, uuid, definition, position, motion, yaw, pitch, headYaw);
     }
